trainable mora-yahi and arbitraryVr
